Admin Implementation

+++ models/engines/mongodb_storage.py
@@ -3,6 +3,12 @@
 Module that abstracts mongodb storage
 """
 import typing as t
+from mongoengine import (
+    connect,
+    disconnect,
+    get_db
+)
+from os import getenv
 from datetime import datetime
 from mongoengine import Document
 
@@ -11,6 +17,10 @@ class MongoDBStorage:
     """
     Class that abstracts mongodb storage
     """
+    def __init__(self):
+        self.db_name = getenv('DB_NAME', 'job_crawler_db')
+        self.db_host = getenv('DB_HOST', 'localhost')
+        self.db_port = int(getenv('DB_PORT', 27017))
 
     def insert(self, collection: t.Type[Document], **kwargs: t.Mapping) -> Document:
         """
@@ -57,15 +67,34 @@ class MongoDBStorage:
         if doc:
             for key, val in kwargs.items():
                 setattr(doc, key, val)
+                if key == 'password':
+                    doc.pre_save()
 
-            self.updated_at = datetime.now()
+            doc.updated_at = datetime.now()
             doc.save()
 
         return self.find(collection, id=id)
+
+    def add_relationship(self, doc: Document, field: str, value: t.Any) -> Document:
+        rel = getattr(doc, field)
+        rel.append(value)
+        doc.save()
+
+    def remove_relationship(self, doc: Document, field: str, value: t.Any) -> Document:
+        rel = getattr(doc, field)
+        rel.remove(value)
+        doc.save()
 
     def delete(self, collection: t.Type[Document], id: str) -> None:
         """
         Delete a document from the collection
         """
         doc = self.find(collection, id=id)
-        doc.delete()
+        if doc:
+            doc.delete()
+
+    def drop_database(self):
+        connect(self.db_name, host=self.db_host, port=self.db_port)
+        db = get_db()
+        db.client.drop_database(db.name)
+        disconnect()
